Enum onnxruntime::error::OrtError

#[non_exhaustive]
pub enum OrtError {

    Library(Error),
    StringConversion(OrtApiError),
    Environment(OrtApiError),
    SessionOptions(OrtApiError),
    Session(OrtApiError),
    Allocator(OrtApiError),
    InOutCount(OrtApiError),
    InputName(OrtApiError),
    GetTypeInfo(OrtApiError),
    CastTypeInfoToTensorInfo(OrtApiError),
    TensorElementType(OrtApiError),
    GetDimensionsCount(OrtApiError),
    GetDimensions(OrtApiError),
    CreateCpuMemoryInfo(OrtApiError),
    CreateTensor(OrtApiError),
    CreateTensorWithData(OrtApiError),
    FillStringTensor(OrtApiError),
    IsTensor(OrtApiError),
    GetTensorTypeAndShape(OrtApiError),
    Run(OrtApiError),
    GetTensorMutableData(OrtApiError),
    DownloadError(OrtDownloadError),
    NonMatchingDimensions(NonMatchingDimensionsError),
    FileDoesNotExists {
        filename: PathBuf,
    },
    NonUtf8Path {
        path: PathBuf,
    },
    CStringNulError(NulError),
    PointerShouldBeNull(String),
    PointerShouldNotBeNull(String),
    InvalidDimensions,
    UndefinedTensorElementType,
    IsTensorCheck,

}

Error type centralizing all possible errors

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Library(Error)

For errors with libloading

StringConversion(OrtApiError)

The C API can message to the caller using a C char * which needs to be converted to Rustโs String. This operation can fail.

Environment(OrtApiError)

An error occurred when creating an ONNXRuntime environment

SessionOptions(OrtApiError)

Error occurred when creating an ONNXRuntime session options

Session(OrtApiError)

Error occurred when creating an ONNXRuntime session

Allocator(OrtApiError)

Error occurred when creating an ONNXRuntime allocator

InOutCount(OrtApiError)

Error occurred when counting ONNXRuntime input or output count

InputName(OrtApiError)

Error occurred when getting ONNXRuntime input name

GetTypeInfo(OrtApiError)

Error occurred when getting ONNXRuntime type information

CastTypeInfoToTensorInfo(OrtApiError)

Error occurred when casting ONNXRuntime type information to tensor information

TensorElementType(OrtApiError)

Error occurred when getting tensor elements type

GetDimensionsCount(OrtApiError)

Error occurred when getting ONNXRuntime dimensions count

GetDimensions(OrtApiError)

Error occurred when getting ONNXRuntime dimensions

CreateCpuMemoryInfo(OrtApiError)

Error occurred when creating CPU memory information

CreateTensor(OrtApiError)

Error occurred when creating ONNXRuntime tensor

CreateTensorWithData(OrtApiError)

Error occurred when creating ONNXRuntime tensor with specific data

FillStringTensor(OrtApiError)

Error occurred when filling a tensor with string data

IsTensor(OrtApiError)

Error occurred when checking if ONNXRuntime tensor was properly initialized

GetTensorTypeAndShape(OrtApiError)

Error occurred when getting tensor type and shape

Run(OrtApiError)

Error occurred when ONNXRuntime inference operation was called

GetTensorMutableData(OrtApiError)

Error occurred when extracting data from an ONNXRuntime tensor into an C array to be used as an ndarray::ArrayView

DownloadError(OrtDownloadError)

Error occurred when downloading a pre-trained ONNX model from the ONNX Model Zoo

NonMatchingDimensions(NonMatchingDimensionsError)

Dimensions of input data and ONNX model loaded from file do not match

FileDoesNotExists

Fields

filename: PathBuf

Path which does not exists

File does not exists

NonUtf8Path

Fields

path: PathBuf

Path with invalid UTF-8

Path is an invalid UTF-8

CStringNulError(NulError)

Attempt to build a Rust CString from a null pointer

PointerShouldBeNull(String)

Ort Pointer should have been null

PointerShouldNotBeNull(String)

Ort pointer should not have been null

InvalidDimensions

ONNXRuntime Model has invalid dimensions

UndefinedTensorElementType

The runtime type was undefined

IsTensorCheck

Error occurred when checking if ONNXRuntime tensor was properly initialized