Trait onnxruntime::Utf8Data

pub trait Utf8Data {
    fn utf8_bytes(&self) -> &[u8] ;
}

Adapter for common Rust string types to Onnx strings.

It should be easy to use both String and &str as TensorElementDataType::String data, but we can’t define an automatic implementation for anything that implements AsRef<str> as it would conflict with the implementations of TypeToTensorElementDataType for primitive numeric types (which might implement AsRef<str> at some point in the future).

Required Methods

fn utf8_bytes(&self) -> &[u8]

Returns the utf8 contents.